move the localizer stuff to its own folder
