+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Drive/Localizer/Localizer.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Subsystems;
+package org.firstinspires.ftc.teamcode.Subsystems.Drive.Localizer;
 
 import org.firstinspires.ftc.teamcode.Subsystems.Drive.Controller.ControllerOutput;
 import org.firstinspires.ftc.teamcode.Subsystems.Drive.MotorGeneric;
